+++ src/Reader/JavaProperties.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * @see       https://github.com/zendframework/zend-config for the canonical source repository
- * @copyright Copyright (c) 2005-2017 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2005-2018 Zend Technologies USA Inc. (http://www.zend.com)
  * @license   https://github.com/zendframework/zend-config/blob/master/LICENSE.md New BSD License
  */
 
@@ -14,12 +14,47 @@ use Zend\Config\Exception;
  */
 class JavaProperties implements ReaderInterface
 {
+    const DELIMITER_DEFAULT = ':';
+    const WHITESPACE_TRIM = true;
+    const WHITESPACE_KEEP = false;
+
     /**
      * Directory of the Java-style properties file
      *
      * @var string
      */
     protected $directory;
+
+    /**
+     * Delimiter for key/value pairs.
+     */
+    private $delimiter;
+
+    /*
+     * Whether or not to trim whitespace from discovered keys and values.
+     *
+     * @var bool
+     */
+    private $trimWhitespace;
+
+    /**
+     * @param string $delimiter Delimiter to use for key/value pairs; defaults
+     *     to self::DELIMITER_DEFAULT (':')
+     * @param bool $trimWhitespace
+     * @throws Exception\InvalidArgumentException for invalid $delimiter values.
+     */
+    public function __construct($delimiter = self::DELIMITER_DEFAULT, $trimWhitespace = self::WHITESPACE_KEEP)
+    {
+        if (! is_string($delimiter) || '' === $delimiter) {
+            throw new Exception\InvalidArgumentException(sprintf(
+                'Invalid delimiter of type "%s"; must be a non-empty string',
+                is_object($delimiter) ? get_class($delimiter) : gettype($delimiter)
+            ));
+        }
+
+        $this->delimiter = $delimiter;
+        $this->trimWhitespace = (bool) $trimWhitespace;
+    }
 
     /**
      * fromFile(): defined by Reader interface.
@@ -99,6 +134,8 @@ class JavaProperties implements ReaderInterface
      */
     protected function parse($string)
     {
+        $delimiter = $this->delimiter;
+        $delimLength = strlen($delimiter);
         $result = [];
         $lines = explode("\n", $string);
         $key = "";
@@ -107,14 +144,15 @@ class JavaProperties implements ReaderInterface
             // Ignore empty lines and commented lines
             if (empty($line)
                || (! $isWaitingOtherLine && strpos($line, "#") === 0)
-               || (! $isWaitingOtherLine && strpos($line, "!") === 0)) {
+               || (! $isWaitingOtherLine && strpos($line, "!") === 0)
+            ) {
                 continue;
             }
 
             // Add a new key-value pair or append value to a previous pair
             if (! $isWaitingOtherLine) {
-                $key = substr($line, 0, strpos($line, ':'));
-                $value = substr($line, strpos($line, ':') + 1, strlen($line));
+                $key = substr($line, 0, strpos($line, $delimiter));
+                $value = substr($line, strpos($line, $delimiter) + $delimLength, strlen($line));
             } else {
                 $value .= $line;
             }
@@ -126,6 +164,11 @@ class JavaProperties implements ReaderInterface
             } else {
                 $isWaitingOtherLine = false;
             }
+
+            $key = $this->trimWhitespace ? trim($key) : $key;
+            $value = $this->trimWhitespace && ! $isWaitingOtherLine
+                ? trim($value)
+                : $value;
 
             $result[$key] = stripslashes($value);
             unset($lines[$i]);